functions/train_pipeline/pyproject.toml
[tool.poetry]
name = "fraud-detection"
version = "0.1.0"
description = ""
authors = ["vitorbeltrao <vitorbeltraoo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
seaborn = "^0.13.2"
imbalanced-learn = "^0.12.2"
pytest = "^8.2.0"
boto3 = "^1.34.106"
python-decouple = "^3.8"
ml-insights = "^1.0.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
autopep8 = "^2.1.0"
pylint = "^3.2.0"
mlflow = "^2.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
